Application/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-bootstrap4==24.3
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
